+++ src/components/section/Header.jsx
@@ -1,17 +1,17 @@
-import { useCallback } from 'react'
-import AppSwitch from '../AppSwitch/index';
+import { useCallback, useState } from 'react'
 import { useTheme } from '../../contexts/theme';
 import Logo from '../header/Logo';
 import Menu from '../header/Menu';
-import Sns from '../header/Sns';
 
 const Header = () => {
-    const { isDarkMode, toggleMode } = useTheme();
     const { theme } = useTheme();
+    const [ isMenuActive, setIsMenuActive ] = useState(false);
 
-    const handleToggle = useCallback(() => {
-      toggleMode();
-    }, [toggleMode]);
+
+
+    const toggleMenu = () => {
+        setIsMenuActive(!isMenuActive);
+    }
 
     const {
         Header: { backgroundColor, textColor, borderColor },
@@ -22,19 +22,12 @@ const Header = () => {
         color: textColor
     };
 
-    const containerStyles = {
-        borderColor: borderColor,
-    };
+
 
     return (
-        <header id='header' role='banner' style={boxStyles}>
-            <Logo />
+        <header id='header' role='banner' style={boxStyles} className={isMenuActive ? 'active' : ''}>
+            <Logo toggleMenu={toggleMenu}/>
             <Menu />
-            <Sns />
-            <div style={containerStyles} className='switch'>
-                <AppSwitch value={isDarkMode} onToggle={handleToggle} ratio={2} />
-                {isDarkMode ? '다크' : '라이트'} 모드
-            </div>
         </header>
     )
 }
